move SqlOmittingQueryOptions to internal package as promised by Javadoc
this change impacts Hibernate Reactive
